优化zlm状态变化事件代码
